+++ src/host_cmd.cc
@@ -799,7 +799,7 @@ void Host_SavegameComment(char *text)
 
 	for (i = 0; i < SAVEGAME_COMMENT_LENGTH; i++)
 		text[i] = ' ';
-	memcpy(text, cl.levelname, min(strlen(cl.levelname), 22)); //johnfitz -- only copy 22 chars.
+	memcpy(text, cl.levelname, min(strlen(cl.levelname), (size_t)22)); //johnfitz -- only copy 22 chars.
 	snprintf(kills, sizeof(kills), "kills:%3i/%3i", cl.stats[STAT_MONSTERS], cl.stats[STAT_TOTALMONSTERS]);
 	memcpy(text + 22, kills, strlen(kills));
 // convert space to _ to make stdio happy
